# docs/node_swagger.yaml
definitions:
...
  big.Int:
    type: object
...
  shared.GenericAPIResponse:
    properties:
      code:
        $ref: '#/definitions/shared.ReturnCode'
      data: {}
      error:
        type: string
    type: object
  shared.ReturnCode:
    enum:
    - successful
    - internal_issue
    - bad_request
    - system_busy
    type: string
    x-enum-varnames:
    - ReturnCodeSuccess
    - ReturnCodeInternalError
    - ReturnCodeRequestError
    - ReturnCodeSystemBusy
  state.List:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    format: int32
    type: integer
    x-enum-varnames:
    - List_inactive
    - List_observer
    - List_waiting
    - List_eligible
    - List_elected
    - List_leaving
    - List_jailed
  state.PeerAccountData:
    properties:
      AccumulatedFees:
        type: integer
      BLSPublicKey:
        items:
          type: integer
        type: array
      ConsecutiveProposerMisses:
        type: integer
      Index:
        type: integer
      LeaderSuccessRate:
        $ref: '#/definitions/state.SignRate'
      List:
        $ref: '#/definitions/state.List'
      NumSelectedInSuccessBlocks:
        type: integer
      OwnerAddress:
        items:
          type: integer
        type: array
      Rating:
        type: integer
      Revoked:
        type: boolean
      RootHash:
        items:
          type: integer
        type: array
      ShardId:
        type: integer
      TempRating:
        type: integer
      TotalLeaderSuccessRate:
        $ref: '#/definitions/state.SignRate'
      TotalValidatorIgnoredSignaturesRate:
        type: integer
      TotalValidatorSuccessRate:
        $ref: '#/definitions/state.SignRate'
      ValidatorIgnoredSignaturesRate:
        type: integer
      ValidatorSuccessRate:
        $ref: '#/definitions/state.SignRate'
    type: object
  state.SignRate:
    properties:
      NumFailure:
        type: integer
      NumSuccess:
        type: integer
    type: object
  state.ValidatorApiResponse:
    properties:
      NumLeaderFailure:
        type: integer
      NumLeaderSuccess:
        type: integer
      NumValidatorFailure:
        type: integer
      NumValidatorIgnoredSignatures:
        type: integer
      NumValidatorSuccess:
        type: integer
      Rating:
        type: number
      RatingModifier:
        type: number
      Revoked:
        type: boolean
      TempRating:
        type: number
      TotalNumLeaderFailure:
        type: integer
      TotalNumLeaderSuccess:
        type: integer
      TotalNumValidatorFailure:
        type: integer
      TotalNumValidatorIgnoredSignatures:
        type: integer
      TotalNumValidatorSuccess:
        type: integer
      ValidatorStatus:
        type: string
    type: object
...
  vm.CallType:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - DirectCall
    - KDATransferAndExecute
    - ExecOnDestByCaller
  vm.KDATransferApi:
    properties:
      tokenName:
        type: string
      tokenNonce:
        type: integer
      tokenType:
        type: integer
      value:
        $ref: '#/definitions/big.Int'
    type: object
  vm.LogEntryApi:
    properties:
      address:
        type: string
      data:
        items:
          items:
            format: int32
            type: integer
          type: array
        type: array
      identifier:
        items:
          type: integer
        type: array
      topics:
        items:
          items:
            format: int32
            type: integer
          type: array
        type: array
    type: object
  vm.OutputAccountApi:
    properties:
      address:
        type: string
      callType:
        $ref: '#/definitions/vm.CallType'
      code:
        items:
          type: integer
        type: array
      codeMetaData:
        items:
          type: integer
        type: array
      outputTransfers:
        items:
          $ref: '#/definitions/vm.OutputTransferApi'
        type: array
      storageUpdates:
        additionalProperties:
          $ref: '#/definitions/vm.StorageUpdateApi'
        type: object
    type: object
  vm.OutputTransferApi:
    properties:
      data:
        items:
          type: integer
        type: array
      receiver:
        type: string
      senderAddress:
        type: string
      tokenInfo:
        $ref: '#/definitions/vm.KDATransferApi'
    type: object
  vm.StorageUpdateApi:
    properties:
      data:
        items:
          type: integer
        type: array
      offset:
        items:
          type: integer
        type: array
    type: object
  vm.VMOutputApi:
    properties:
      deletedAccounts:
        items:
          items:
            format: int32
            type: integer
          type: array
        type: array
      gasRemaining:
        type: integer
      logs:
        items:
          $ref: '#/definitions/vm.LogEntryApi'
        type: array
      outputAccounts:
        additionalProperties:
          $ref: '#/definitions/vm.OutputAccountApi'
        type: object
      returnCode:
        type: string
      returnData:
        items:
          items:
            format: int32
            type: integer
          type: array
        type: array
      returnMessage:
        type: string
    type: object
  vm.VMValueRequest:
    properties:
      args:
        items:
          type: string
        type: array
      caller:
        type: string
      funcName:
        type: string
      sameScState:
        type: boolean
      scAddress:
        type: string
      shouldBeSynced:
        type: boolean
      value:
        additionalProperties:
          format: int64
          type: integer
        type: object
    type: object
info:
  contact: {}
  description: This is a sample of the KleverChain Node Server.
  title: KleverChain Node
  version: "1.0"
paths:
...
  /validator/peers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  properties:
                    peers:
                      items:
                        $ref: '#/definitions/state.PeerAccountData'
                      type: array
                  type: object
              type: object
        "400":
          description: some error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: will return the peers
      tags:
      - Validator
  /validator/statistics:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  properties:
                    statistics:
                      additionalProperties:
                        $ref: '#/definitions/state.ValidatorApiResponse'
                      type: object
                  type: object
              type: object
        "400":
          description: some error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: will return the validation statistics for all validators
      tags:
      - Validator
  /vm/hex:
    post:
      parameters:
      - description: vm request data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/vm.VMValueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  properties:
                    data:
                      type: string
                  type: object
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns the data as bytes, hex-encoded
      tags:
      - VM
  /vm/int:
    post:
      parameters:
      - description: vm request data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/vm.VMValueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  properties:
                    data:
                      type: string
                  type: object
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns the data as big int
      tags:
      - VM
  /vm/query:
    post:
      parameters:
      - description: vm request data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/vm.VMValueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  properties:
                    data:
                      $ref: '#/definitions/vm.VMOutputApi'
                  type: object
              type: object
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns the VM output of a smart contract query
      tags:
      - VM
  /vm/string:
    post:
      parameters:
      - description: vm request data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/vm.VMValueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  properties:
                    data:
                      type: string
                  type: object
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns the data as string
      tags:
      - VM
swagger: "2.0"